requirements.txt
Flask==1.1.1
jikanpy==4.2.2
Pillow==8.1.0
uWSGI==2.0.17.1